server/requirements.txt
colorlog==3.1.2
Flask_Cors==3.0.3
Flask==0.12.2
requests==2.18.4
numpy==1.14.0
spacy==2.0.7
PyYAML==3.12
pyopenssl==17.5.0